+++ frontend/app/components/home/hero-section.tsx
@@ -1,6 +1,5 @@
 import { ArrowRight } from "lucide-react";
 import Link from "next/link";
-import * as React from "react";
 
 import { SafeSwapLogo } from "@/app/components/shared/safe-swap-logo";
 import { Button } from "@/app/components/ui/button";
