+++ frontend/app/components/products/Images.tsx
@@ -1,5 +1,4 @@
 import Image from "next/image";
-import React from "react";
 
 interface ImageProps {
 	images: {
